# gh-110874: http.cookiejar.deepvalues improve performance