--- a/Lib/http/cookiejar.py
+++ b/Lib/http/cookiejar.py
@@ -1227,12 +1227,17 @@ class DefaultCookiePolicy(CookiePolicy):
 def deepvalues(mapping):
     """Iterates over nested mapping, depth-first"""
     for obj in tuple(mapping.values()):
-        mapping = True
-        try:
-            obj.items
-        except AttributeError:
-            mapping = False
-        if mapping:
+        # mapping = True
+        # try:
+        #     obj.items
+        # except AttributeError:
+        #     mapping = False
+        # if mapping:
+        #     yield from deepvalues(obj)
+        # else:
+        #     yield obj
+
+        if hasattr(obj, "items"):
             yield from deepvalues(obj)
         else:
             yield obj