--- a/Lib/http/cookiejar.py
+++ b/Lib/http/cookiejar.py
@@ -1227,14 +1227,9 @@ class DefaultCookiePolicy(CookiePolicy):
 def deepvalues(mapping):
     """Iterates over nested mapping, depth-first"""
     for obj in tuple(mapping.values()):
-        mapping = True
-        try:
-            obj.items
-        except AttributeError:
-            mapping = False
-        if mapping:
+        if hasattr(obj, 'items'):
             yield from deepvalues(obj)
-        else:
+    else:
             yield obj