--- a/Misc/NEWS.d/next/Core and Builtins/2023-10-14-13-44-20.gh-issue-110874.EdP8y6.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2023-10-14-13-44-20.gh-issue-110874.EdP8y6.rst
@@ -1,0 +1,1 @@
+Rewrite :func:`http.cookiejar.deepvalues`  improve performance. Rewrite by Tunglies.