@Eclips4 2023-10-14
Hello!
Have you made any performance measurements? If so, please, share the results.