@Tunglies 2023-10-14
> Hello! Have you made any performance measurements? If so, please, share the results.

Sure!
Here is my performance measurements code based on httpx.
```
import io
import pstats
import cProfile
import httpx

merge_cookies_times = 1_000_000

pr = cProfile.Profile()
pr.enable()

httpx.Client(cookies={"default":"none"})._merge_cookies(cookies={f"{k}":f"{v}" for k,v in zip(range(merge_cookies_times), range(merge_cookies_times))})

pr.disable()
s = io.StringIO()
ps = (pstats.Stats(pr, stream=s)
      .strip_dirs()
      .print_callers(".deepvalues")
      .sort_stats("tottime"))
ps.print_stats(10)
print(s.getvalue())
```
Before
```
   Random listing order was used
   List reduced from 225 to 1 due to restriction <'.deepvalues'>

Function                       was called by...
                                   ncalls  tottime  cumtime
cookiejar.py:1227(deepvalues)  <-       2    0.000    0.000  _models.py:1024(__init__)
                                  1000001    0.276    1.371  _models.py:1147(update)
                                2000006/1000003    1.095    1.095  cookiejar.py:1227(deepvalues)


         16002280 function calls (14002219 primitive calls) in 12.741 seconds

   Ordered by: internal time
   List reduced from 225 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  2000002    2.880    0.000    3.433    0.000 cookiejar.py:1678(set_cookie)
  1000001    2.714    0.000    8.164    0.000 _models.py:1057(set)
  1000001    1.546    0.000    1.979    0.000 copy.py:66(copy)
  1000001    1.466    0.000    3.557    0.000 cookiejar.py:761(__init__)
3000009/1000003    1.371    0.000    1.371    0.000 cookiejar.py:1227(deepvalues)
        1    0.510    0.510    0.510    0.510 perf_test.py:11(<dictcomp>)
        1    0.487    0.487   12.179   12.179 _models.py:1147(update)
        3    0.444    0.148    8.608    2.869 _models.py:1024(__init__)
  2000003    0.331    0.000    0.331    0.000 {method 'acquire' of '_thread.RLock' objects}
  1000001    0.310    0.000    0.310    0.000 {method 'copy' of 'dict' objects}
```
Update
```
   Random listing order was used
   List reduced from 225 to 1 due to restriction <'.deepvalues'>

Function                       was called by...
                                   ncalls  tottime  cumtime
cookiejar.py:1227(deepvalues)  <-       2    0.000    0.000  _models.py:1024(__init__)
                                  1000001    0.223    1.108  _models.py:1147(update)
                                2000006/1000003    0.885    0.885  cookiejar.py:1227(deepvalues)


         16002280 function calls (14002219 primitive calls) in 10.599 seconds

   Ordered by: internal time
   List reduced from 225 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  2000002    2.418    0.000    2.883    0.000 cookiejar.py:1679(set_cookie)
  1000001    2.225    0.000    6.772    0.000 _models.py:1057(set)
  1000001    1.278    0.000    1.596    0.000 copy.py:66(copy)
  1000001    1.238    0.000    2.929    0.000 cookiejar.py:761(__init__)
3000009/1000003    1.108    0.000    1.108    0.000 cookiejar.py:1227(deepvalues)
        1    0.516    0.516    0.516    0.516 perf_test.py:11(<dictcomp>)
        3    0.371    0.124    7.142    2.381 _models.py:1024(__init__)
        1    0.370    0.370   10.030   10.030 _models.py:1147(update)
  2000003    0.276    0.000    0.276    0.000 {method 'acquire' of '_thread.RLock' objects}
  1000001    0.214    0.000    0.214    0.000 {method 'copy' of 'dict' objects}
```