@gaogaotiantian 2023-10-14
Hmm, your profiling data seems a bit off to me. It looks like ALL of your functions were faster, even `{method 'acquire' of '_thread.RLock' objects}`. Are you using the same enviroment/python for the two tests? Did you make sure the ONLY difference between two runs were the lines you've changed?

Also, cProfile itself is normally used for profiling, not benchmarking. To prove the perf improvement, I think you need `timeit` and a more self-contained example (not with `httpx`). It's not easy to tell from the code change that the perf improvement is significant (it's a micro-optimization).