@Tunglies 2023-10-15
> Hmm, your profiling data seems a bit off to me. It looks like ALL of your functions were faster, even `{method 'acquire' of '_thread.RLock' objects}`. Are you using the same enviroment/python for the two tests? Did you make sure the ONLY difference between two runs were the lines you've changed?
> 
> Also, cProfile itself is normally used for profiling, not benchmarking. To prove the perf improvement, I think you need `timeit` and a more self-contained example (not with `httpx`). It's not easy to tell from the code change that the perf improvement is significant (it's a micro-optimization).

I tried rewrite code below.
```python
import timeit
from Lib.http.cookiejar import Cookie, CookieJar


def cookie_data(key, value):
    cookie_data = {
        "version": 0,
        "name": key,
                "value": value,
                "port": None,
                "port_specified": False,
                "domain": "",
                "domain_specified": False,
                "domain_initial_dot": False,
                "path": "/",
                "path_specified": True,
                "secure": False,
                "expires": None,
                "discard": True,
                "comment": None,
                "comment_url": None,
                "rest": {"HttpOnly": None},
                "rfc2109": False,
    }
    return cookie_data


def test_code():
    cookie_jar = CookieJar()
    cookie = Cookie(**cookie_data("Key", "Value"))
    cookie_jar.set_cookie(cookie)
    for i in range(1_000_000):
        _cookie_data = Cookie(**cookie_data(f"{i}", f"{i}"))
        cookie_jar.set_cookie(_cookie_data)


if __name__ == "__main__":
    execution_time = timeit.timeit(test_code, number=5)
    print(execution_time)
```
And here is results.
Unchanged `12.56033199999365` 
Changed `12.346485399990343`